requirements.txt
pytest
allure-pytest
playwright
pytest-playwright
PyMySQL
